requirements.txt
torch==2.4.1
torchvision==0.19.1
PILLOW==10.4.0
numpy==1.26.4
tensorflow==2.17.0